Update `git`, and `prerelease`.
`git` and `prerelease` use .NET 10 (release candidate 1), which was
broken on Hydra for a while. I didn't want people to suffer building it
themselves, so I opted to wait until it was successfully built again.

`prerelease` is currently pinned to the commit after the actual
Pre-Release tag on GitHub. This is due to the build erroring,
but the developer of Aaru made another commit after the
Pre-Release tag that fixes that exact issue so I'm using that instead.